fix: requested changes applied
