+++ internal/mapper/user_v2/user_v2_resource_mapper.go
@@ -5,22 +5,15 @@ import (
 
 	mapper "github.com/conduktor/terraform-provider-conduktor/internal/mapper"
 	"github.com/conduktor/terraform-provider-conduktor/internal/model"
-	schemaUtils "github.com/conduktor/terraform-provider-conduktor/internal/schema"
-	schema "github.com/conduktor/terraform-provider-conduktor/internal/schema/resource_user_v2"
-	helpers "github.com/conduktor/terraform-provider-conduktor/internal/test"
+	schema "github.com/conduktor/terraform-provider-conduktor/internal/schema"
+	users "github.com/conduktor/terraform-provider-conduktor/internal/schema/resource_user_v2"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
-func TFToInternalModel(ctx context.Context, r *schema.UserV2Model) (model.UserConsoleResource, error) {
-	if r.Spec.Permissions.IsNull() {
-		tflog.Debug(ctx, "ListValue externalGroups is null")
-	} else if r.Spec.Permissions.IsUnknown() {
-		tflog.Debug(ctx, "ListValue externalGroups is unknown")
-	}
-	permissions, err := helpers.SetValueToPermissionArray(ctx, "users", r.Spec.Permissions)
+func TFToInternalModel(ctx context.Context, r *users.UserV2Model) (model.UserConsoleResource, error) {
+	permissions, err := schema.SetValueToPermissionArray(ctx, schema.USERS, r.Spec.Permissions)
 	if err != nil {
 		return model.UserConsoleResource{}, err
 	}
@@ -35,29 +28,29 @@ func TFToInternalModel(ctx context.Context, r *schema.UserV2Model) (model.UserCo
 	), nil
 }
 
-func InternalModelToTerraform(ctx context.Context, r *model.UserConsoleResource) (schema.UserV2Model, error) {
-	permissionsList, err := helpers.PermissionArrayToSetValue(ctx, "users", r.Spec.Permissions)
+func InternalModelToTerraform(ctx context.Context, r *model.UserConsoleResource) (users.UserV2Model, error) {
+	permissionsList, err := schema.PermissionArrayToSetValue(ctx, schema.USERS, r.Spec.Permissions)
 	if err != nil {
-		return schema.UserV2Model{}, err
+		return users.UserV2Model{}, err
 	}
 
-	specValue, diag := schema.NewSpecValue(
+	specValue, diag := users.NewSpecValue(
 		map[string]attr.Type{
 			"firstname":   basetypes.StringType{},
 			"lastname":    basetypes.StringType{},
 			"permissions": permissionsList.Type(ctx),
 		},
 		map[string]attr.Value{
-			"firstname":   schemaUtils.NewStringValue(r.Spec.FirstName),
-			"lastname":    schemaUtils.NewStringValue(r.Spec.LastName),
+			"firstname":   schema.NewStringValue(r.Spec.FirstName),
+			"lastname":    schema.NewStringValue(r.Spec.LastName),
 			"permissions": permissionsList,
 		},
 	)
 	if diag.HasError() {
-		return schema.UserV2Model{}, mapper.WrapDiagError(diag, "spec", mapper.IntoTerraform)
+		return users.UserV2Model{}, mapper.WrapDiagError(diag, "spec", mapper.IntoTerraform)
 	}
 
-	return schema.UserV2Model{
+	return users.UserV2Model{
 		Name: types.StringValue(r.Metadata.Name),
 		Spec: specValue,
 	}, nil
